Fix indexing in extract_annotation_from()
json_obj['annotations'] contains the annotations in chronological
order, i.e., the first annotation is the oldest, and the last
annotation is the latest.
In order to recognize any changes to the annotations and provide
Timewarrior with the latest annotation, the last annotation with
index -1 should be returned instead of the first one.